// apps/u3/src/services/types/activity.ts
type Json = any;

export type Stream = {
  streamId: string;
  did: string;
  network: string;
  indexingTime: number;
  familyOrApp: string | null;
  type: string;
  from: string;
  tags: string[];
  status: string;
  hash: string;
  schema: string;
  model?: string;
  anchorStatus: string;
  commitIds: string[];
  content: Json;
  metadata: Json;
  domain?: string;
};

export enum Network {
  MAINNET = 'MAINNET',
  TESTNET = 'TESTNET',
}